+++ src/dao/sql/favorites.ts
@@ -27,7 +27,7 @@ SELECT
   ak.duration AS duration,
   ak.created_at AS created_at,
   ak.modified_at AS modified_at,
-  count(ak.kid) OVER() AS count
+  count(ak.kid) OVER()::integer AS count
   FROM all_karas AS ak
   INNER JOIN favorites AS f ON f.fk_kid = ak.kid
   WHERE fk_login = :username
