+++ src/dao/sql/kara.ts
@@ -90,7 +90,7 @@ export const getAllKaras = (filterClauses: string[], lang: LangClause, typeClaus
   ak.repo AS repo,
   ak.tag_names AS tag_names,
   ak.tid AS tid,
-  count(ak.kid) OVER() AS count
+  count(ak.kid) OVER()::integer AS count
 FROM all_karas AS ak
 LEFT OUTER JOIN kara_serie AS ks_main ON ks_main.fk_kid = ak.kid
 LEFT OUTER JOIN serie_lang AS sl_main ON sl_main.fk_sid = ks_main.fk_sid AND sl_main.lang = ${lang.main}
